Initial support for features in cargo_repository
This adds as many target as needed to support all features
combinaison if the feature is indeed available (i.e. all
features / deps are available). Each target is prefixed with
`@a+b` where `a` and `b` are activated features.
`default` denotes the default feature.

The support is still missing support for adding dependencies / declaring
feature dependencies.

Some crate overabuse features and generating all combination is `O(n!)` where
n is the number of allowed features. For winapi, which has 360 features, this
is about 10^690 targets ¯\_(ツ)_/¯. To avoid that we limit the number of
features before we trigger all the target. A better approach would be to use
feature flag but they are not available at the moment.